[mplfe]Internal commit msg:     [mplfeC] clean up pclint warning, part 3     solve 1764 warnings [mapleall]Internal commit msg:     !907 Preserve comparison signedness when doing LFTR     [mplcg] fix conditional branch range calculate for alignment     !909 Added load/store pair of q-registers. Split out --tailcall option in mplcg     [mapleall] enhance deal with phi in vrp     [mapleall] Prediction bugfix

+++ src/mplfe/common/src/feir_var_reg.cpp
@@ -66,8 +66,8 @@ bool FEIRVarReg::EqualsToImpl(const std::unique_ptr<FEIRVar> &var) const {
   return ptrVarReg->regNum == regNum;
 }
 
-size_t FEIRVarReg::HashImpl() const {
-  return std::hash<uint32>{}(regNum);
+uint32 FEIRVarReg::HashImpl() const {
+  return static_cast<uint32>(std::hash<uint32>{}(regNum));
 }
 
 // ========== FEIRVarAccumulator ==========
